feat: 메인페이지 레이어 계층 분리 (#39)
