+++ src/pages/Main/index.tsx
@@ -1,18 +1,20 @@
 import styled from "styled-components";
 
 import { calcResponsive } from "@common/styles/theme";
-import MyPlannerCardList from "@components/cardList/MyPlannerCardList";
-import StartPlannerCard from "@components/card/StartPlannerCard";
 import Header from "@components/header/Header";
+import StartPlannerCardContainer from "./containers/StartPlannerCardContainer";
+import MyPlannerCardListContainer from "./containers/MyPlannerCardListContainer";
+import TokenCookieManager from "@components/debug/TokenCookieManager";
 
 const Main = () => {
   return (
     <>
       <Header />
       <Container>
-        <MyPlannerCardList items={mock} onEmptyCardClick={() => {}} />
-        <StartPlannerCard onClickPlanner={() => {}} onSubmit={() => {}} />
+        <MyPlannerCardListContainer />
+        <StartPlannerCardContainer />
       </Container>
+      <TokenCookieManager />
     </>
   );
 };
@@ -32,36 +34,3 @@ const Container = styled.div`
 `;
 
 export default Main;
-
-const mock = [
-  {
-    country: "KOREA",
-    startDate: "2024.11.24",
-    endDate: "2024.11.30",
-    onClick: () => {},
-  },
-  {
-    country: "JAPAN",
-    startDate: "2025.01.01",
-    endDate: "2025.01.07",
-    onClick: () => {},
-  },
-  {
-    country: "CHINA",
-    startDate: "2025.02.01",
-    endDate: "2025.02.07",
-    onClick: () => {},
-  },
-  {
-    country: "JAPAN",
-    startDate: "2025.03.01",
-    endDate: "2025.03.07",
-    onClick: () => {},
-  },
-  {
-    country: "KOREA",
-    startDate: "2025.11.24",
-    endDate: "2025.11.30",
-    onClick: () => {},
-  },
-];
