+++ src/pages/Modal/RoomCreateModal.tsx
@@ -1,0 +1,5 @@
+const RoomCreateModal = () => {
+  return <div>RoomCreateModal</div>;
+};
+
+export default RoomCreateModal;
