+++ src/pages/Modal/RoomEnterModal.tsx
@@ -1,0 +1,5 @@
+const RoomEnterModal = () => {
+  return <div>RoomEnterModal</div>;
+};
+
+export default RoomEnterModal;
