+++ src/pages/Modal/index.ts
@@ -1,0 +1,2 @@
+export { default as RoomCreateModal } from "./RoomCreateModal";
+export { default as RoomEnterModal } from "./RoomEnterModal";
